Refactor Cryptor class
